Migliorie varie

+++ src/state/result_state.rs
@@ -4,6 +4,6 @@ use borsh::{BorshSerialize, BorshDeserialize};
 
 #[derive(BorshSerialize, BorshDeserialize)]
 pub struct ResultState {
-    pub results: HashMap<String, f64>,
+    pub results: HashMap<String, f32>,
     pub number_of_votes: i64
 }
